define Automaton interface locally
In Go we prefer to define interfaces where we consume them.
Since the Automaton is now a core part of how scorch allows
one to iterate a term dictionary, this definition belongs here.

As a side-effect, this removes dependency on the vellum package
from the API (although both Bleve and Zap implementation code
continues to use the vellum package).